+++ src/syntax/mod.rs
@@ -1,0 +1,2 @@
+mod lexer;
+mod tokens;
